Blacklist

+++ MultiDriveSync/RemoteFileSynchronizer.cs
@@ -12,12 +12,14 @@ namespace MultiDriveSync
         private readonly IGoogleDriveClient googleDriveClient;
         private readonly Dictionary<string, string> parentPathsById;
         private readonly string rootId;
+        private readonly MultiDriveSyncService _service;
 
-        public RemoteFileSynchronizer(IGoogleDriveClient googleDriveClient, string localRootPath, string rootId)
+        public RemoteFileSynchronizer(IGoogleDriveClient googleDriveClient, string localRootPath, string rootId, MultiDriveSyncService service)
         {
             this.googleDriveClient = googleDriveClient ?? throw new ArgumentNullException(nameof(googleDriveClient));
             parentPathsById = new Dictionary<string, string>();
             this.rootId = rootId;
+            _service = service;
             parentPathsById[rootId] = localRootPath;
         }
 
@@ -43,18 +45,38 @@ namespace MultiDriveSync
                     switch (change.ChangeContentType)
                     {
                         case ChangeContentType.Folder when change.ChangeType == ChangeType.CreatedOrUpdated:
+                            if (_service.BlackList.Contains(change.Id))
+                            {
+                                _service.BlackList.Remove(change.Id);
+                                continue;
+                            }
                             CreateOrUpdateFolder(change.ParentId, change.Id, change.Name);
                             break;
 
                         case ChangeContentType.Folder when change.ChangeType == ChangeType.Deleted:
+                            if (_service.BlackList.Contains(change.Id))
+                            {
+                                _service.BlackList.Remove(change.Id);
+                                continue;
+                            }
                             DeleteFolder(change.Id);
                             break;
 
                         case ChangeContentType.File when change.ChangeType == ChangeType.CreatedOrUpdated:
+                            if (_service.BlackList.Contains(change.Id))
+                            {
+                                _service.BlackList.Remove(change.Id);
+                                continue;
+                            }
                             await CreateOrUpdateFile(change.ParentId, change.Id, change.Name);
                             break;
 
                         case ChangeContentType.File when change.ChangeType == ChangeType.Deleted:
+                            if (_service.BlackList.Contains(change.Id))
+                            {
+                                _service.BlackList.Remove(change.Id);
+                                continue;
+                            }
                             DeleteFile(change.ParentId, change.Name);
                             break;
